Move external functions and dataclasses to telemetry_utils
